Update file 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,3 @@
-#include "function_pointers.h"
 #include <stdio.h>
 /**
  * array_iterator - give a parameter on each element of an array
@@ -6,7 +5,7 @@
  * @size_t size: size of array
  * @action: give the accion
  */
-void array_iterator(int *array, int size, void (*action)(int))
+void array_iterator(int *array, size_t size, void (*action)(int))
 {
 	int i;
